+++ tests/checks/ending_blank_line.rs
@@ -15,7 +15,7 @@ fn correct_files() {
 
         let expected_output = check_output(&[(".env", &[])]);
 
-        testdir.test_command_success_with_args(args, expected_output);
+        testdir.test_command_success_with_args(with_default_args(args), expected_output);
     }
 }
 
@@ -42,6 +42,6 @@ fn incorrect_files() {
             .as_str()],
         )]);
 
-        testdir.test_command_fail_with_args(args, expected_output);
+        testdir.test_command_fail_with_args(with_default_args(args), expected_output);
     }
 }