+++ tests/common/test_dir.rs
@@ -65,28 +65,6 @@ impl TestDir {
             .expect("convert directory to &str")
     }
 
-    /// Run the default CLI binary in this TestDir and check it succeeds.
-    ///
-    /// This method removes the TestDir when command has finished.
-    pub fn test_command_success<T>(self, expected_output: T)
-    where
-        T: Into<String>,
-    {
-        let args: &[&str; 0] = &[];
-        self.test_command_success_with_args(args, expected_output);
-    }
-
-    /// Run the default CLI binary in this TestDir and check it fails.
-    ///
-    /// This method removes the TestDir when command has finished.
-    pub fn test_command_fail<T>(self, expected_output: T)
-    where
-        T: Into<String>,
-    {
-        let args: &[&str; 0] = &[];
-        self.test_command_fail_with_args(args, expected_output);
-    }
-
     /// Run the default CLI binary, with command line arguments,
     /// in this TestDir and check it succeeds.
     ///